+++ lib/src/callbacks.dart
@@ -1,4 +1,3 @@
-
 part of flutter_naver_map;
 
 typedef void MapCreateCallback(NaverMapController controller);